Add Pair module and make necessary changes
Added hundreds of lines of code to implement the Pair module which is the foundation for liquidity pools on Uniswap. Adjustments were also made to the other modules and files to accommodate this addition. Other changes include the conversion of core constants to static lazy consts and improvement in the readability and functionality of the code.