Add method for finding best trade with fixed input
A new method `best_trade_exact_in` has been added to the trade entity. This method returns the top trades for a specific input token amount and an output token. It considers valid routes with a defined maximum number of hops. Additional imports relevant to the implementation of this function have also been included.